+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Notable additions, fixes, or breaking changes to the Freeplay SDK.
 
+## [0.5.5] - 2025-10-27
+
+- New examples around multimodal images as output
+- Get test cases and output_message which may be more than just text content. 'output' is now deprecated.
+
 ## [0.5.4] - 2025-10-17
 
 - Add explicit tool span logging